fix(client): split ChatRoom out of ChatPage so useChat cannot run before the auth guard
useChat() connects a socket on mount, so with RequireAuth and the hook in the
same component, an anonymous visitor to /chat opened (and the server
rejected) a socket before being redirected. ChatPage is now a guard shell
that renders ChatRoom only once RequireAuth lets a signed-in viewer through;
ChatRoom carries the hook and all rendering. Tests moved with it, plus a new
guard test asserting useChat is never called for an unauthenticated visitor.
