+++ apps/client/src/pages/ChatPage.test.tsx
@@ -2,49 +2,26 @@
 // never runs — every client test file wires jest-dom and cleanup itself, as
 // apps/client/src/components/patterns/CommentForm.test.tsx does.
 import '@testing-library/jest-dom/vitest'
-import type { ChatMessage } from '@blog/zod-shared'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { cleanup, fireEvent, render, screen } from '@testing-library/react'
+import { cleanup, render } from '@testing-library/react'
 import { MemoryRouter } from 'react-router'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import { queryKeys } from '../lib/query-client.js'
-import type { ChatStatus, ChatUser } from '../hooks/use-chat.js'
 
-// Mocked so each test can drive ChatPage through a specific hook state
-// directly, instead of exercising the real socket.
+// Mocked as a spy, not stubbed with real state, because the property under
+// test is whether this hook gets CALLED at all — useChat() connects a socket
+// on mount, so calling it for an anonymous visitor opens (and the server
+// immediately rejects) a socket for nothing, waking the realtime service's
+// free-tier instance on every anonymous hit to /chat. This is the regression
+// the ChatPage/ChatRoom split exists to prevent.
 vi.mock('../hooks/use-chat.js', () => ({ useChat: vi.fn() }))
 
 const { useChat } = await import('../hooks/use-chat.js')
 const { ChatPage } = await import('./ChatPage.js')
 
-type ChatState = {
-  messages: ChatMessage[]
-  online: ChatUser[]
-  typingUsers: string[]
-  status: ChatStatus
-  send: ReturnType<typeof vi.fn>
-  setTyping: ReturnType<typeof vi.fn>
-}
-
-function mockChat(overrides: Partial<ChatState> = {}): ChatState {
-  const state: ChatState = {
-    messages: [],
-    online: [],
-    typingUsers: [],
-    status: 'connected',
-    send: vi.fn(),
-    setTyping: vi.fn(),
-    ...overrides,
-  }
-  vi.mocked(useChat).mockReturnValue(state)
-  return state
-}
-
-function renderChatPage() {
+function renderChatPage(me: { id: string; username: string } | null) {
   const client = new QueryClient({ defaultOptions: { queries: { retry: false } } })
-  // RequireAuth is UX only, not under test here — pre-seed a signed-in viewer
-  // so it resolves past the loading/redirect states immediately.
-  client.setQueryData(queryKeys.me, { id: 'u1', username: 'reader' })
+  client.setQueryData(queryKeys.me, me)
   return render(
     <MemoryRouter>
       <QueryClientProvider client={client}>
@@ -54,128 +31,24 @@ function renderChatPage() {
   )
 }
 
-describe('ChatPage status line', () => {
+describe('ChatPage guard', () => {
   afterEach(() => cleanup())
 
-  // The status line is the entire mitigation for the ~60s cold start — a
-  // silent page during that window is indistinguishable from a broken one.
-  it('tells the reader the socket is connecting', () => {
-    mockChat({ status: 'connecting' })
-    renderChatPage()
-    expect(screen.getByText('Connecting…')).toBeInTheDocument()
+  it('never calls useChat — and so never opens a socket — for an unauthenticated visitor', () => {
+    renderChatPage(null)
+    expect(useChat).not.toHaveBeenCalled()
   })
 
-  it('tells the reader the socket dropped and is retrying', () => {
-    mockChat({ status: 'reconnecting' })
-    renderChatPage()
-    expect(screen.getByText('Reconnecting…')).toBeInTheDocument()
-  })
-
-  it('tells the reader the connection failed and how to recover', () => {
-    mockChat({ status: 'failed' })
-    renderChatPage()
-    expect(screen.getByText('Could not connect. Reload to try again.')).toBeInTheDocument()
-  })
-
-  it('shows no status line once connected', () => {
-    mockChat({ status: 'connected' })
-    renderChatPage()
-    expect(screen.queryByText('Connecting…')).not.toBeInTheDocument()
-    expect(screen.queryByText('Reconnecting…')).not.toBeInTheDocument()
-    expect(screen.queryByText('Could not connect. Reload to try again.')).not.toBeInTheDocument()
-  })
-})
-
-describe('ChatPage send flow', () => {
-  afterEach(() => cleanup())
-
-  it('sends the typed body, then clears the draft and stops the typing signal', () => {
-    const state = mockChat({ status: 'connected' })
-    renderChatPage()
-
-    fireEvent.change(screen.getByLabelText('Message'), { target: { value: 'hello room' } })
-    fireEvent.click(screen.getByText('Send'))
-
-    expect(state.send).toHaveBeenCalledWith('hello room')
-    expect(screen.getByLabelText('Message')).toHaveValue('')
-    expect(state.setTyping).toHaveBeenCalledWith(false)
-  })
-})
-
-describe('ChatPage send button gating', () => {
-  afterEach(() => cleanup())
-
-  it('disables Send when the draft is empty', () => {
-    mockChat({ status: 'connected' })
-    renderChatPage()
-    expect(screen.getByText('Send')).toBeDisabled()
-  })
-
-  it('disables Send when the draft is only whitespace', () => {
-    mockChat({ status: 'connected' })
-    renderChatPage()
-    fireEvent.change(screen.getByLabelText('Message'), { target: { value: '   ' } })
-    expect(screen.getByText('Send')).toBeDisabled()
-  })
-
-  it('disables Send while not connected, even with a non-empty draft', () => {
-    mockChat({ status: 'reconnecting' })
-    renderChatPage()
-    fireEvent.change(screen.getByLabelText('Message'), { target: { value: 'hello' } })
-    expect(screen.getByText('Send')).toBeDisabled()
-  })
-
-  it('enables Send once there is a non-empty draft and the socket is connected', () => {
-    mockChat({ status: 'connected' })
-    renderChatPage()
-    fireEvent.change(screen.getByLabelText('Message'), { target: { value: 'hello' } })
-    expect(screen.getByText('Send')).not.toBeDisabled()
-  })
-})
-
-describe('ChatPage message input gating', () => {
-  afterEach(() => cleanup())
-
-  it('disables the input once the connection has failed', () => {
-    mockChat({ status: 'failed' })
-    renderChatPage()
-    expect(screen.getByLabelText('Message')).toBeDisabled()
-  })
-
-  it('leaves the input enabled while connected', () => {
-    mockChat({ status: 'connected' })
-    renderChatPage()
-    expect(screen.getByLabelText('Message')).not.toBeDisabled()
-  })
-})
-
-describe('ChatPage messages and presence', () => {
-  afterEach(() => cleanup())
-
-  it('renders a message with its author and body', () => {
-    mockChat({
-      messages: [
-        {
-          id: 'm1',
-          body: 'hello room',
-          author: { id: 'u2', username: 'abe' },
-          sentAt: '2026-01-01T00:00:00.000Z',
-        },
-      ],
+  it('calls useChat once RequireAuth has let a signed-in viewer through', () => {
+    vi.mocked(useChat).mockReturnValue({
+      messages: [],
+      online: [],
+      typingUsers: [],
+      status: 'connected',
+      send: vi.fn(),
+      setTyping: vi.fn(),
     })
-    renderChatPage()
-    expect(screen.getByText('abe')).toBeInTheDocument()
-    expect(screen.getByText(/hello room/)).toBeInTheDocument()
-  })
-
-  it('reflects the online roster size', () => {
-    mockChat({
-      online: [
-        { id: 'u1', username: 'reader' },
-        { id: 'u2', username: 'abe' },
-      ],
-    })
-    renderChatPage()
-    expect(screen.getByText('2 online')).toBeInTheDocument()
+    renderChatPage({ id: 'u1', username: 'reader' })
+    expect(useChat).toHaveBeenCalled()
   })
 })
